bugfix: Exclude newer LSP4j, which is built on JDK 11
It seems the CI started failing because of out bump, but that was not picked up aside from windows tests weirdly.

Metals recently updated to newest LSP4j, though I think nothing new was added and no changes were done to the presentation compiler so we should be ok to downgrade locally: https://github.com/scalameta/metals/pull/6126/files

The update was mostly done so that we done fall behind in the LSP standard, but this might become a problem in the future here. Do we ever plan do drop supporting JDK 8?

[Cherry-picked 790bee903c3ac2ffc2e7254535e03cfc08f3e023]