Parser simple expression error recovery change from `null` to `???` (#19103)
Previously, simpleExpr was recovered as `Literal(Constant(null))` which
led to some errors in interactive.

Type inference in Scala 3 works on whole chain, thus type vars were
inferred as union type of `Null` because of this very reason. Recovering
such errors as `unimplementedExpr` which has a type of `Nothing`, solves
the issue.
[Cherry-picked 98184f15234426384fad476e5587c7aa2aa2a215][modified]