Make opaque type refinements of inline proxy objects abstract tycons
Change the rhs of the added refinements from `TypeAlias`es to `RealTypeBounds`s.
This allows the opaque types to remain abstract type constructors,
which is necessary for the `MatchTypeCasePattern.AbstractTypeConstructor` logic.

In particular, an AppliedType where the tycon was a reference to the refinement,
used to dealias before proceeding with the comparison of the tycons,
which is a requirement of the aforementioned `AbstractTypeConstructor`
case of the MatchReducer.

Fixes #20427
Alternative to #20457