Fix isomorphism tests of `AndOrType`s under non-empty `BinderPairs`
Before the changes, when comparing two `HKTypeLambda` result types,
the list of binding pairs was lost when entering comparison of `AndOrType`s,
which caused the `equals` to fail, and hence prevented `hash-consing`.

Even though, `M1` and `M2` in pos/i20858-min should still conform to one-another,
we entered a deep-subtype comparison because of the order in which the TypeComparer
does dealiasing of AppliedTypes, comparison of MatchCases, and AndTypes.