Restore forgotten fatal-warnings tests
In daeee3544a7933736240d1b79ebe81d699d74b0d, support for tests in the directory
`tests/pos-special/fatal-warnings` was dropped in favor of magic `//>` comments
in the regular `tests/pos` directory, but a few tests were forgotten in the
original directory and were thus not run. This commit moves them to the
appropriate directory with the correct magic comment.

[Cherry-picked 0ee804f5e5ad4e8712bb11c6a9f1bb172ef7586b][modified]