Avoid the TypeVar.inst trap
`tvar.inst` gives the _permanent_ instance of a type variable `tvar`. Even if `tvar.isInstantiated` is true
its `inst` can still be NoType. This is a trap that caused a regression in the code of glb. This commit fixes
the regression and introduces different names that will hopefully avoid the trap in the future.

Fixes #20154

[Cherry-picked 2863a29f9e6c7685bd0bf632fa02b322f59ce99f][modified]