Fix #20856: Serialize `Waiting` and `Evaluating` as if `null`.
This strategy ensures the "serializability" condition of parallel
programs--not to be confused with the data being `java.io.Serializable`.
Indeed, if thread A is evaluating the lazy val while thread B attempts
to serialize its owner object, there is also an alternative schedule
where thread B serializes the owner object *before* A starts evaluating
the lazy val. Therefore, forcing B to see the non-evaluating state is
correct.

[Cherry-picked e2427533c563cb5ca708dbcaa51823fcc30102db]