Disambiguate by owner score before considering further implicit arguments
This makes slick-migration-api-example work and makes the original scala-uri.scala fail.
See the comment in neg/scala-uri.scala for why this is so. It also needs a change in the
givens in dotty.tools.reporting.Formatting.

The motivation for honoring owner score over the others is that it is often used for explicit
prioritization, so we should take it into account more than other secondary criteria.