A tweak to type improvement
When we replace Nothing by a fresh type variable, we should not accidentally
instantiate that type variable to Any in case it is still undetermined. We achieve
this by giving the type variable a slightly disguised version of Nothing which
makes the compiler believe it has a lower bound.

Fixes #21275

[Cherry-picked 053c8f9cd2c382481d147d90e37c6c8c327e2882]