Test overloading changes and reported warning messages
Note that in tests/neg/multiparamlist-overload-3.7 Test2 we now
get an error in both Parts as intended. But they are, however, different ones:
Part1 is a TypeMismatch Error, whereas Part2 is a NoMatchingOverload Error.

This is due to the fact that `resolveOverloaded` will first
find the candidate alternatives by considering only the 1st parameter list
and commit early if there is a single one, e.g. Test2.Part1.
If not, we recursively continue with the found alternatives and
recompute the candidate alternatives based on the 2nd parameter list,
which may rule out all of them, and hence lead to a different message.