Fix stale symbol errors connected to lazyvals
The actual stale symbol error, while recent, seems correct to me.
There is no reason why an earlier phase (Mixins) should see denotation
from a later phase (LazyVals) - if the opposite is true (like with the fix
here) the bringForward method is able to recover with a NoDenotation,
which is what we want.