Fix pkg obj prefix of opaque tp ext meth
There is use of `makePackageObjPrefixExplicit` within `accessibleType`,
which is called on the result of findRef in typedIdent.  But in
`tryExtension` it's no such use.  We could fix it in the usage of the
results in `tryExtension`, but I thought perhaps we could fix it for all
call sites, by handling it within findRef.