Drop special handling in recheckApplication
In recheckApplication we implemented a rule that computed the capture set of an
application `f(a)` by being the "better" of either the capture set of `f` union
capture set of `a` or the capture set of the result type of `f(a)`. This relies
on capture monotinicity, which we try to get away from. Also, it's semantically
dubious if the type of the argument `a` is a type variable that can be instantiated
to a capturing type as in the following example:
```scala
trait Iterator[+A]
  ...
  def flatMap[B](f: A => IterableOnce[B]^): Iterator[B]^{this, f}
```
Here, we account for every capability in `IterableOnce[B]^` already in `f`. But
those capabilities could also come from `A` if `A` is instantiated in the
actual function argument to a capturing type.

There was extensive breakage once the special handling was dropped. One example is
the `flatMap` definition above. We leave that as a potential soundness hold for now,
but the right way to fix `flatMap` would be by adding a capture set variable:
```scala
def flatMap[B, C^](f: A => IterableOnce[B]^{C^}): Iterator[B]^{this, f, C^}
```
The problem is that this currently cannot be done in a way that allows flatMap to be called
as before, passing a single type argument for `B`. We'd have to change the system to either
allow curried type parameters or optional type parameters for capture sets.

Another issue is that now more reach capabilities are registered as used by the enclosing method.
An example is lazylist-exceptions.scala, which has been moved to pending. Here, the use is spurious
because it happens inside an anonymous class creation on the right hand side of the method.
With refined use checking, the use would not propagate to the method, so the reach capability should
not be leaking. But to get there we need to implement refined use checking first.