Standardize how previously experimental features are handled
If a feature was previously experimental and is now standard, we change
any tests for that feature to be only dependent on the source version where
the feature was standardized. Language imports in old source versions
will no longer enable the feature. (And these language imports
also come with a deprecation message).

If a feature was previously experimental and is now dropped, the feature
becomes unavailable also in old versions.

The motivation to do it this way is to insist that experimental features
are ephemeral. We should not be able to rely on an experimental feature
forever in an old version.

This commit implements this policy for fewerBraces and clauseInterleaving.
Two implemented extensions (relaxedExtensionImports, betterMatchTypeExtractors)
already implemented it before.