SIP 61 - remove tasty hack by not "telescoping" forwarders
Now, all forwarders directly call the unrolled method.
This means that there is no need to resolve an invisible method,
so the "hack" using TERMREFdirect is no longer needed.

This increases bytecode size but reduces stack depth.

Also removed scala.annotation.internal.UnrolledForwarder,
as it is only needed for the TASTy hack.