add failing test: demo transparent inline can not see forwarders at typer
At typer, when transparent inline methods are expanded, unroll forwarders
are not yet generated.

So if we define forwarders for the first time, in the same compilation
unit that inlines a call to the forwarder, then it will fail.

Perhaps we can detect this and suspend the unit, otherwise it would seem
unreasonable complexity to generate forwarders in typer.