Make `Applications#resolveMapped` preserve annotations
Overloading may create a temporary symbol via `Applications#resolveMapped`, but
before this commit, this symbol did not carry the annotations from the original
symbol. This meant in particular that `isInlineable` would always return false
for them. This matters because during the course of overloading resolution we
might call `ProtoTypes.Compatibility#constrainResult` which special-cases
transparent inline methods.

Fixes a regression in Monocle introduced in the previous commit.

wip

[Cherry-picked f7259cf77e053b00068bd5dca124e9a3013489dd]