Avoid even more false negatives in overload pruning
The changes two commits ago were not enough to handle i21410b.scala because we
end up checking:

    Tuple.Map[WildcardType(...), List] <: (List[Int], List[String])

which fails because a match type with a wildcard argument apparently only gets
reduced when the match type case is not parameterized.

To handle this more generally we use AvoidWildcardsMap to remove wildcards from
the result type, but since we want to prevent false negatives we start with
`variance = -1` to get a lower-bound instead of an upper-bound.

[Cherry-picked 32ac2e62f9b18aecf0064236a7ce6c60a1bbf579]