Downgrade referenceVersion to 3.6.0 and add notes how referenceVersion should be set (#21813)
The main branch should always use a compiler with experimental tasty -
it's required for non_boostrapped tests to pass. It's required because
we cannot consume stable version of tasty from it's experimental
subversion (it's assumed that stable version > experimental version for
the same major/minor pair)

We pick 3.6.0 (released by mistake during 3.6.0-RC1) because that's the
last non-stable version pushed to Maven. Typically we would use the last
RC version

[test_non_bootstrapped]