Add implied captures in function types
This is an attempt to fix the problem explified in the `delayedRunops*.scala` tests.
We can treat it as a baseline that fixes the immediate problem of the interaction of
reach capabilities and type variables. There might be better ways to do this by using
a proper adapation rule for function types instead of adding implied captures post-hoc.