Expand value references to packages to their underlying package objects
A package object can be seen as the facade of a package. For instance, it is the logical
place where we want to write doc comments that explain a package.

So far references to packages cannot be used as values. But if the package has a package
object, it would make sense to allow the package reference with the meaning that it refers
to this object. For instance, let's say we have

```scala
package a
object b
```
Of course, we can use `a.b` as a value. But if we change that to
```scala
package a
package object b
```
we can't anymore. This PR changes that so that we still allow a reference `a.b`
as a value to mean the package object. Due to the way package objects are encoded
the `a.b` reference expands to `a.b.package`.