Drop reference to implicit conversion from named to regular tuples
The NamedTuple spec mentioned that `.toTuple` was inserted implicitly to convert
from a named to a regular tuple. But this was not implemented. The sentiment is
that this presents an opportunity to get more experience with the feature before
deciding whether the conversion is actually beneficial. For now, one can implement
the conversion by hand if desired, like this:
```scala
given [N <: Tuple, V <: Tuple] => Conversion[NamedTuple.NamedTuple[N, V], V] = _.toTuple
```