Fix inline reduction for CaseDef guards with asInstanceOf (#22305)
In Inliner.scala
we add asInstanceOf to references to private inline methods to make sure
we later are able to know which method is referenced (if e.g. we inline
out of the scope where that method would be visible). This added
asInstanceOf caused issue when inlining CaseDef guards, as instead of a
simple constant literal we get an Inlined node with an added binding,
like this:
```scala
{
  val A_this: A = A_this.asInstanceOf[A]
  true:Boolean
}
```
We fix that by just unpacking that Inlined node (and we do not need that
binding for constant literals, so we can just ignore it).
[Cherry-picked 1448123ce79ac96cc1de25ee84fd9a1fb063f3ba]