Fix coverage serialization when encountering macro suspension  (#22303)
Fixes #22247

The fix is simple, as we mainly move the coverage object to a global
ContextBase object, which persists it between runs. Initially I thought
that appending the newly generated coverage indices would be enough, but
if the macro suspends after the InstrumentCoverage phase runs, we end up
with duplicate indices. For that reason, when generating indexes for a
compilation unit, we also remove the previously generated ones for the
same compilation unit.

To support having multiple scala files compiled in the coverage tests I
had to slightly adjust the suite. While doing that, I noticed that some
check files for run tests were ignored, as they were incorrectly named.
I added an assertion that throws when `.check` do not exist and renamed
the files appropriately (having to add some additional ones as well).