Fix mapping of annotations
We previously did not do an (expensive) TreeTypeMap on an annotation if the mapped versions
of all types of subtrees of the annotation tree were =:= to the original types. But it turns
out this is too coarse. In the test we have capture set variables where we intend to map a
TypeRef to a TypeParamRef but the two were considered as =:= because of the bounds they had.
So no mapping took place.

We now use `eql` instead of =:=, which is structural comparison with `eq` for references to
corresponding binders.